renamed files to follow convention
